# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Learning Summit API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type Track = (typeof Track)[keyof typeof Track];

export const Track = {
  lower: "lower",
  middle: "middle",
  upper: "upper",
  all: "all",
  required_all: "required_all",
  teachers: "teachers",
} as const;

export type Role = (typeof Role)[keyof typeof Role];

export const Role = {
  student: "student",
  faculty: "faculty",
  organizer: "organizer",
  admin: "admin",
} as const;

export interface User {
  id: string;
  email: string;
  firstName?: string | null;
  lastName?: string | null;
  imageUrl?: string | null;
  role: Role;
  division: Track;
  createdAt: string;
}

export interface UpdateMeBody {
  role?: Role;
  division?: Track;
}

export interface Speaker {
  name: string;
  title?: string | null;
  bio?: string | null;
}

export interface Session {
  id: string;
  title: string;
  description: string;
  location: string;
  room: string;
  track: Track;
  mandatory: boolean;
  capacity: number;
  registeredCount: number;
  startsAt: string;
  endsAt: string;
  organizers: string[];
  speakers: Speaker[];
  tags: string[];
  isRegistered: boolean;
}

export type SessionDetail = Session & {
  attendees: User[];
};

export interface CreateSessionBody {
  title: string;
  description: string;
  location: string;
  room: string;
  track: Track;
  mandatory?: boolean;
  capacity: number;
  startsAt: string;
  endsAt: string;
  organizers?: string[];
  speakers?: Speaker[];
  tags?: string[];
}

export interface UpdateSessionBody {
  title?: string;
  description?: string;
  location?: string;
  room?: string;
  track?: Track;
  mandatory?: boolean;
  capacity?: number;
  startsAt?: string;
  endsAt?: string;
  organizers?: string[];
  speakers?: Speaker[];
  tags?: string[];
}

export interface Registration {
  id: string;
  sessionId: string;
  userId: string;
  createdAt: string;
}

export type CheckInMethod = (typeof CheckInMethod)[keyof typeof CheckInMethod];

export const CheckInMethod = {
  qr: "qr",
  room: "room",
} as const;

export interface CheckInBody {
  method: CheckInMethod;
  code?: string | null;
}

export interface Attendance {
  id: string;
  sessionId: string;
  userId: string;
  method: CheckInMethod;
  checkedInAt: string;
}

export interface AttendanceWithUser {
  id: string;
  sessionId: string;
  userId: string;
  method: CheckInMethod;
  checkedInAt: string;
  user: User;
}

export interface Feedback {
  id: string;
  sessionId: string;
  userId: string;
  rating: number;
  comment?: string | null;
  createdAt: string;
  userName?: string | null;
}

export interface SubmitFeedbackBody {
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment?: string | null;
}

export interface FeedbackSummary {
  averageRating: number;
  count: number;
  items: Feedback[];
}

export type SessionFeedbackAggregateRecentCommentsItem = {
  comment: string;
  userName?: string | null;
  rating: number;
  createdAt: string;
};

export interface SessionFeedbackAggregate {
  sessionId: string;
  sessionTitle: string;
  track: Track;
  startsAt: string;
  averageRating: number;
  count: number;
  recentComments: SessionFeedbackAggregateRecentCommentsItem[];
}

export interface InvitedStudent {
  id: string;
  email: string;
  firstName?: string | null;
  lastName?: string | null;
  division: Track;
  registered: boolean;
  registeredAt?: string | null;
  invitedAt: string;
}

export interface ImportInvitedStudentRow {
  email: string;
  firstName?: string | null;
  lastName?: string | null;
  division?: Track;
}

export interface ImportInvitedStudentsBody {
  rows: ImportInvitedStudentRow[];
}

export interface ImportInvitedStudentsResult {
  inserted: number;
  skipped: number;
}

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
  blocked: "blocked",
} as const;

export interface Task {
  id: string;
  title: string;
  description?: string | null;
  assigneeId: string;
  assigneeName?: string | null;
  assigneeEmail?: string | null;
  status: TaskStatus;
  dueAt?: string | null;
  createdAt: string;
  completedAt?: string | null;
}

export interface CreateTaskBody {
  title: string;
  description?: string | null;
  assigneeId: string;
  dueAt?: string | null;
  status?: TaskStatus;
}

export interface UpdateTaskBody {
  title?: string;
  description?: string | null;
  assigneeId?: string;
  dueAt?: string | null;
  status?: TaskStatus;
}

export interface TaskExportRow {
  id: string;
  title: string;
  description?: string | null;
  assigneeName?: string | null;
  assigneeEmail?: string | null;
  status: string;
  dueAt?: string | null;
  completedAt?: string | null;
  createdAt: string;
}

export interface SessionExportRow {
  id: string;
  title: string;
  track: string;
  room: string;
  capacity: number;
  registeredCount: number;
  attendedCount: number;
  startsAt: string;
}

export interface TaskExport {
  tasks: TaskExportRow[];
  sessions: SessionExportRow[];
}

export type FormFieldType = (typeof FormFieldType)[keyof typeof FormFieldType];

export const FormFieldType = {
  text: "text",
  textarea: "textarea",
  number: "number",
  select: "select",
  checkbox: "checkbox",
} as const;

export interface FormField {
  key: string;
  label: string;
  type: FormFieldType;
  required: boolean;
  options?: string[];
}

export interface Form {
  id: string;
  title: string;
  description?: string | null;
  sessionId?: string | null;
  creatorId: string;
  creatorName?: string | null;
  responseCount: number;
  createdAt: string;
}

export type FormDetail = Form & {
  fields: FormField[];
};

export interface CreateFormBody {
  title: string;
  description?: string | null;
  sessionId?: string | null;
  fields: FormField[];
}

export type FormResponseValues = { [key: string]: unknown };

export interface FormResponse {
  id: string;
  formId: string;
  userId: string;
  userName?: string | null;
  values: FormResponseValues;
  createdAt: string;
}

export type SubmitFormResponseBodyValues = { [key: string]: unknown };

export interface SubmitFormResponseBody {
  values: SubmitFormResponseBodyValues;
}

export type NotificationLevel =
  (typeof NotificationLevel)[keyof typeof NotificationLevel];

export const NotificationLevel = {
  info: "info",
  warning: "warning",
  emergency: "emergency",
} as const;

export interface Notification {
  id: string;
  title: string;
  body: string;
  level: NotificationLevel;
  createdAt: string;
  creatorId: string;
  creatorName?: string | null;
}

export interface CreateNotificationBody {
  title: string;
  body: string;
  level: NotificationLevel;
}

export interface TrackBreakdown {
  track: Track;
  sessionCount: number;
  registeredCount: number;
}

export interface DashboardSummary {
  totalSessions: number;
  totalAttendees: number;
  totalCheckIns: number;
  upcomingCount: number;
  fullSessionsCount: number;
  openTaskCount: number;
  averageRating: number;
  unreadNotifications: Notification[];
  trackBreakdown: TrackBreakdown[];
  nextSessions: Session[];
}

export type ListSessionsParams = {
  track?: Track;
};