# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Learning Summit API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sessions/{id}/feedback:
    get:
      operationId: listSessionFeedback
      tags: [feedback]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Feedback list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackSummary"
    post:
      operationId: submitFeedback
      tags: [feedback]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitFeedbackBody"
      responses:
        "201":
          description: Feedback created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feedback"
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /forms:
    get:
      operationId: listForms
      tags: [forms]
      responses:
        "200":
          description: Forms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Form"
    post:
      operationId: createForm
      tags: [forms]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFormBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Form"
  /forms/{id}:
    get:
      operationId: getForm
      tags: [forms]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Form
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormDetail"
  /forms/{id}/responses:
    post:
      operationId: submitFormResponse
      tags: [forms]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitFormResponseBody"
      responses:
        "201":
          description: Response submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormResponse"
    get:
      operationId: listFormResponses
      tags: [forms]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Responses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FormResponse"
...
  /students:
    get:
      operationId: listInvitedStudents
      tags: [students]
      summary: List invited students with registration status
      responses:
        "200":
          description: Invited students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InvitedStudent"
    post:
      operationId: importInvitedStudents
      tags: [students]
      summary: Bulk-create invited students from a parsed roster
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportInvitedStudentsBody"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportInvitedStudentsResult"
  /students/{id}:
    delete:
      operationId: deleteInvitedStudent
      tags: [students]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed
...
components:
  schemas:
...
    Track:
      type: string
      enum: [lower, middle, upper, all, required_all, teachers]
...
    Feedback:
      type: object
      properties:
        id: { type: string }
        sessionId: { type: string }
        userId: { type: string }
        rating: { type: integer }
        comment: { type: string, nullable: true }
        createdAt: { type: string }
        userName: { type: string, nullable: true }
      required: [id, sessionId, userId, rating, createdAt]
    SubmitFeedbackBody:
      type: object
      properties:
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string, nullable: true }
      required: [rating]
    FeedbackSummary:
      type: object
      properties:
        averageRating: { type: number }
        count: { type: integer }
        items:
          type: array
          items: { $ref: "#/components/schemas/Feedback" }
      required: [averageRating, count, items]
...
    InvitedStudent:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        firstName: { type: string, nullable: true }
        lastName: { type: string, nullable: true }
        division: { $ref: "#/components/schemas/Track" }
        registered: { type: boolean }
        registeredAt: { type: string, nullable: true }
        invitedAt: { type: string }
      required: [id, email, division, registered, invitedAt]
    ImportInvitedStudentRow:
      type: object
      properties:
        email: { type: string }
        firstName: { type: string, nullable: true }
        lastName: { type: string, nullable: true }
        division: { $ref: "#/components/schemas/Track" }
      required: [email]
    ImportInvitedStudentsBody:
      type: object
      properties:
        rows:
          type: array
          items: { $ref: "#/components/schemas/ImportInvitedStudentRow" }
      required: [rows]
    ImportInvitedStudentsResult:
      type: object
      properties:
        inserted: { type: integer }
        skipped: { type: integer }
      required: [inserted, skipped]
    TaskStatus:
      type: string
      enum: [todo, in_progress, done, blocked]
    Task:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        assigneeId: { type: string }
        assigneeName: { type: string, nullable: true }
        assigneeEmail: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/TaskStatus" }
        dueAt: { type: string, nullable: true }
        createdAt: { type: string }
        completedAt: { type: string, nullable: true }
      required: [id, title, assigneeId, status, createdAt]
    CreateTaskBody:
      type: object
      properties:
        title: { type: string }
        description: { type: string, nullable: true }
        assigneeId: { type: string }
        dueAt: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/TaskStatus" }
      required: [title, assigneeId]
    UpdateTaskBody:
      type: object
      properties:
        title: { type: string }
        description: { type: string, nullable: true }
        assigneeId: { type: string }
        dueAt: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/TaskStatus" }
...
    FormFieldType:
      type: string
      enum: [text, textarea, number, select, checkbox]
    FormField:
      type: object
      properties:
        key: { type: string }
        label: { type: string }
        type: { $ref: "#/components/schemas/FormFieldType" }
        required: { type: boolean }
        options: { type: array, items: { type: string } }
      required: [key, label, type, required]
    Form:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        sessionId: { type: string, nullable: true }
        creatorId: { type: string }
        creatorName: { type: string, nullable: true }
        responseCount: { type: integer }
        createdAt: { type: string }
      required: [id, title, creatorId, responseCount, createdAt]
    FormDetail:
      allOf:
        - $ref: "#/components/schemas/Form"
        - type: object
          properties:
            fields:
              type: array
              items: { $ref: "#/components/schemas/FormField" }
          required: [fields]
    CreateFormBody:
      type: object
      properties:
        title: { type: string }
        description: { type: string, nullable: true }
        sessionId: { type: string, nullable: true }
        fields:
          type: array
          items: { $ref: "#/components/schemas/FormField" }
      required: [title, fields]
    FormResponse:
      type: object
      properties:
        id: { type: string }
        formId: { type: string }
        userId: { type: string }
        userName: { type: string, nullable: true }
        values:
          type: object
          additionalProperties: true
        createdAt: { type: string }
      required: [id, formId, userId, values, createdAt]
    SubmitFormResponseBody:
      type: object
      properties:
        values:
          type: object
          additionalProperties: true
      required: [values]
...