# lib/api-zod/src/generated/types/importInvitedStudentsResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Learning Summit API
 * OpenAPI spec version: 0.1.0
 */

export interface ImportInvitedStudentsResult {
  inserted: number;
  skipped: number;
}