# lib/api-zod/src/generated/types/track.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Learning Summit API
 * OpenAPI spec version: 0.1.0
 */

export type Track = (typeof Track)[keyof typeof Track];

export const Track = {
  lower: "lower",
  middle: "middle",
  upper: "upper",
  all: "all",
  required_all: "required_all",
  teachers: "teachers",
} as const;